+++ temba/flows/tests.py
@@ -21,6 +21,7 @@ from django.utils.encoding import force_text
 
 from temba.airtime.models import AirtimeTransfer
 from temba.api.models import Resthook, WebHookEvent, WebHookResult
+from temba.archives.models import Archive
 from temba.channels.models import Channel, ChannelEvent
 from temba.contacts.models import TEL_SCHEME, URN, Contact, ContactField, ContactGroup, ContactURN
 from temba.ivr.models import IVRCall
@@ -36,6 +37,7 @@ from temba.tests import (
     matchers,
     skip_if_no_flowserver,
 )
+from temba.tests.s3 import MockS3Client
 from temba.triggers.models import Trigger
 from temba.ussd.models import USSDSession
 from temba.utils.dates import datetime_to_str
@@ -1080,7 +1082,7 @@ class FlowTest(TembaTest):
         )
 
         # test without msgs or unresponded
-        with self.assertNumQueries(35):
+        with self.assertNumQueries(36):
             workbook = self.export_flow_results(flow, include_msgs=False, responded_only=True)
 
         tz = self.org.timezone
@@ -1167,6 +1169,13 @@ class FlowTest(TembaTest):
         self.assertExcelRow(
             sheet_runs,
             1,
+            [contact3_run1.contact.uuid, "+250788123456", "Norbert", "", contact3_run1.created_on, "", "", "", ""],
+            tz,
+        )
+
+        self.assertExcelRow(
+            sheet_runs,
+            2,
             [
                 contact1_run1.contact.uuid,
                 "+250788382382",
@@ -1183,7 +1192,7 @@ class FlowTest(TembaTest):
 
         self.assertExcelRow(
             sheet_runs,
-            2,
+            3,
             [
                 contact2_run1.contact.uuid,
                 "+250788383383",
@@ -1200,14 +1209,14 @@ class FlowTest(TembaTest):
 
         self.assertExcelRow(
             sheet_runs,
-            3,
-            [contact3_run1.contact.uuid, "+250788123456", "Norbert", "", contact3_run1.created_on, "", "", "", ""],
+            4,
+            [contact2_run2.contact.uuid, "+250788383383", "Nic", "", contact2_run2.created_on, "", "", "", ""],
             tz,
         )
 
         self.assertExcelRow(
             sheet_runs,
-            4,
+            5,
             [
                 contact1_run2.contact.uuid,
                 "+250788382382",
@@ -1222,13 +1231,6 @@ class FlowTest(TembaTest):
             tz,
         )
 
-        self.assertExcelRow(
-            sheet_runs,
-            5,
-            [contact2_run2.contact.uuid, "+250788383383", "Nic", "", contact2_run2.created_on, "", "", "", ""],
-            tz,
-        )
-
         # check messages sheet...
         self.assertEqual(len(list(sheet_msgs.rows)), 14)  # header + 13 messages
         self.assertEqual(len(list(sheet_msgs.columns)), 7)
@@ -1238,10 +1240,25 @@ class FlowTest(TembaTest):
         contact1_out1 = contact1_run1.get_messages().get(text="What is your favorite color?")
         contact1_out2 = contact1_run1.get_messages().get(text="That is a funny color. Try again.")
         contact1_out3 = contact1_run1.get_messages().get(text__startswith="I love orange too")
+        contact3_out1 = contact3_run1.get_messages().get(text="What is your favorite color?")
 
         self.assertExcelRow(
             sheet_msgs,
             1,
+            [
+                self.contact3.uuid,
+                "+250788123456",
+                "Norbert",
+                contact3_out1.created_on,
+                "OUT",
+                "What is your favorite color?",
+                "Test Channel",
+            ],
+            tz,
+        )
+        self.assertExcelRow(
+            sheet_msgs,
+            2,
             [
                 contact1_out1.contact.uuid,
                 "+250788382382",
@@ -1255,7 +1272,7 @@ class FlowTest(TembaTest):
         )
         self.assertExcelRow(
             sheet_msgs,
-            2,
+            3,
             [
                 contact1_in1.contact.uuid,
                 "+250788382382",
@@ -1269,7 +1286,7 @@ class FlowTest(TembaTest):
         )
         self.assertExcelRow(
             sheet_msgs,
-            3,
+            4,
             [
                 contact1_out2.contact.uuid,
                 "+250788382382",
@@ -1283,7 +1300,7 @@ class FlowTest(TembaTest):
         )
         self.assertExcelRow(
             sheet_msgs,
-            4,
+            5,
             [
                 contact1_in2.contact.uuid,
                 "+250788382382",
@@ -1297,7 +1314,7 @@ class FlowTest(TembaTest):
         )
         self.assertExcelRow(
             sheet_msgs,
-            5,
+            6,
             [
                 contact1_out3.contact.uuid,
                 "+250788382382",
@@ -1312,7 +1329,7 @@ class FlowTest(TembaTest):
         )
 
         # test without msgs or unresponded
-        with self.assertNumQueries(40):
+        with self.assertNumQueries(41):
             workbook = self.export_flow_results(self.flow, include_msgs=False, responded_only=True)
 
         tz = self.org.timezone
@@ -1375,7 +1392,7 @@ class FlowTest(TembaTest):
         age = ContactField.get_or_create(self.org, self.admin, "age", "Age")
         self.contact.set_field(self.admin, "age", 36)
 
-        with self.assertNumQueries(41):
+        with self.assertNumQueries(42):
             workbook = self.export_flow_results(
                 self.flow,
                 include_msgs=False,
@@ -1481,26 +1498,6 @@ class FlowTest(TembaTest):
 
         sheet_runs, sheet_msgs = workbook.worksheets
 
-        # check runs sheet...
-        self.assertEqual(len(list(sheet_runs.rows)), 2)  # header + 1 runs
-        self.assertEqual(len(list(sheet_runs.columns)), 9)
-
-        self.assertExcelRow(
-            sheet_runs,
-            0,
-            [
-                "Contact UUID",
-                "URN",
-                "Name",
-                "Groups",
-                "Started",
-                "Exited",
-                "color (Category) - Color Flow",
-                "color (Value) - Color Flow",
-                "color (Text) - Color Flow",
-            ],
-        )
-
         self.assertExcelRow(
             sheet_runs,
             1,
@@ -1515,6 +1512,105 @@ class FlowTest(TembaTest):
                 "ngertin.",
                 "ngertin.",
             ],
+            tz,
+        )
+
+    def test_export_results_from_archives(self):
+        contact1_run, contact2_run, contact3_run = self.flow.start([], [self.contact, self.contact2, self.contact3])
+        contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="green")
+        Flow.find_and_handle(contact1_in1)
+        contact2_in1 = self.create_msg(direction=INCOMING, contact=self.contact2, text="blue")
+        Flow.find_and_handle(contact2_in1)
+
+        # and a run for a different flow
+        flow2 = self.get_flow("favorites")
+        contact2_other_flow, = flow2.start([], [self.contact2])
+
+        contact1_run.refresh_from_db()
+        contact2_run.refresh_from_db()
+        contact3_run.refresh_from_db()
+
+        # archive runs for contacts 1 and 2
+        Archive.objects.create(
+            org=self.org,
+            archive_type=Archive.TYPE_FLOWRUN,
+            size=10,
+            hash=uuid4().hex,
+            url="http://test-bucket.aws.com/archive1.jsonl.gz",
+            record_count=3,
+            start_date=timezone.now().date(),
+            period="D",
+            build_time=23425,
+        )
+        mock_s3 = MockS3Client()
+        mock_s3.put_jsonl(
+            "test-bucket",
+            "archive1.jsonl.gz",
+            [contact1_run.as_archive_json(), contact2_run.as_archive_json(), contact2_other_flow.as_archive_json()],
+        )
+
+        contact1_run.release()
+        contact2_run.release()
+
+        # create an archive earlier than our flow created date so we check that it isn't included
+        Archive.objects.create(
+            org=self.org,
+            archive_type=Archive.TYPE_FLOWRUN,
+            size=10,
+            hash=uuid4().hex,
+            url="http://test-bucket.aws.com/archive2.jsonl.gz",
+            record_count=1,
+            start_date=timezone.now().date() - timedelta(days=2),
+            period="D",
+            build_time=5678,
+        )
+        mock_s3.put_jsonl("test-bucket", "archive2.jsonl.gz", [contact2_run.as_archive_json()])
+
+        with patch("temba.archives.models.Archive.s3_client", return_value=mock_s3):
+            workbook = self.export_flow_results(self.flow)
+
+        tz = self.org.timezone
+        sheet_runs, sheet_msgs = workbook.worksheets
+
+        # check runs sheet...
+        self.assertEqual(len(list(sheet_runs.rows)), 4)  # header + 3 runs
+
+        self.assertExcelRow(
+            sheet_runs,
+            1,
+            [
+                contact1_run.contact.uuid,
+                "+250788382382",
+                "Eric",
+                "",
+                contact1_run.created_on,
+                "",
+                "Other",
+                "green",
+                "green",
+            ],
+            tz,
+        )
+        self.assertExcelRow(
+            sheet_runs,
+            2,
+            [
+                contact2_run.contact.uuid,
+                "+250788383383",
+                "Nic",
+                "",
+                contact2_run.created_on,
+                contact2_run.exited_on,
+                "Blue",
+                "blue",
+                "blue",
+            ],
+            tz,
+        )
+        self.assertExcelRow(
+            sheet_runs,
+            3,
+            [contact3_run.contact.uuid, "+250788123456", "Norbert", "", contact3_run.created_on, "", "", "", ""],
             tz,
         )
 
@@ -4000,8 +4096,8 @@ class ActionTest(TembaTest):
         self.assertIsNone(replies.filter(contact_urn__path="stephen").first())
         self.assertIsNotNone(replies.filter(contact_urn__path="+12078778899").first())
 
-        Broadcast.objects.all().delete()
-        Msg.objects.all().delete()
+        self.release(Broadcast.objects.all())
+        self.release(Msg.objects.all())
 
         msg = self.create_msg(direction=INCOMING, contact=contact, text="Green is my favorite")
         run = FlowRun.create(self.flow, self.contact)
@@ -4037,8 +4133,9 @@ class ActionTest(TembaTest):
             reply_msg.attachments, ["image/jpeg:https://%s/%s" % (settings.AWS_BUCKET_DOMAIN, "path/to/media.jpg")]
         )
 
-        Broadcast.objects.all().delete()
-        Msg.objects.all().delete()
+        self.release(Broadcast.objects.all())
+        self.release(Msg.objects.all())
+
         msg = self.create_msg(direction=INCOMING, contact=self.contact, text="Green is my favorite")
 
         action_json = action.as_json()
@@ -4758,15 +4855,14 @@ class ActionTest(TembaTest):
 
         self.execute_action(action, run, msg)
 
-        # should do nothing
+        # should do nothing and not log an error
         self.assertEqual(dynamic_group.contacts.count(), 0)
+        self.assertEqual(ActionLog.objects.filter(level="E").count(), 0)
 
         # tho if contact is a test contact, log as error
         self.execute_action(action, test_run, test_msg)
-
         self.assertEqual(dynamic_group.contacts.count(), 0)
-
-        self.assertEqual(ActionLog.objects.filter(level="E").count(), 2)
+        self.assertEqual(ActionLog.objects.filter(level="E").count(), 1)
 
         group1 = self.create_group("Flow Group 1", [])
         group2 = self.create_group("Flow Group 2", [])
@@ -5067,6 +5163,22 @@ class FlowRunTest(TembaTest):
         )
 
         self.assertTrue(FlowRun.objects.get(contact=self.contact).is_interrupted())
+
+    def test_run_release(self):
+        run = FlowRun.create(self.flow, self.contact)
+
+        # give our run some webhook data
+        event = WebHookEvent.objects.create(
+            org=self.org, run=run, channel=self.channel, created_by=self.admin, modified_by=self.admin
+        )
+
+        WebHookResult.objects.create(
+            event=event, status_code=200, url="", created_by=self.admin, modified_by=self.admin
+        )
+
+        # our run go bye bye
+        run.release()
+        self.assertFalse(FlowRun.objects.filter(id=run.id).exists())
 
 
 class FlowLabelTest(FlowFileTest):
@@ -9932,7 +10044,7 @@ class OrderingTest(FlowFileTest):
             # again, only one send_msg
             self.assertEqual(mock_send_msg.call_count, 1)
 
-        Msg.objects.all().delete()
+        self.releaseMessages()
 
         # try with multiple contacts
         with patch("temba.channels.tasks.send_msg_task", wraps=send_msg_task) as mock_send_msg:
@@ -10173,7 +10285,7 @@ class TimeoutTest(FlowFileTest):
 
         # ok, now let's try with a timeout
         self.releaseRuns()
-        Msg.objects.all().delete()
+        self.releaseMessages()
 
         # start the flow
         flow.start([], [self.contact])